Pk-720

+++ src/Kooco.Pikachu.Application.Contracts/GroupBuys/GroupBuyCreateDto.cs
@@ -202,7 +202,7 @@ namespace Kooco.Pikachu.GroupBuys
         public string? CustomerInformationDescription { get; set; }
         public string? ExchangePolicyDescription { get; set; }
         public bool IsEnterprise { get; set; }
-        public decimal? FreeShippingThreshold { get; set; }
+        public int? FreeShippingThreshold { get; set; }
         public List<string>? ShippingMethodList { get; set; }
         public string? SelfPickupDeliveryTime { get; set; }
         public string? BlackCatDeliveryTime { get; set; }
